Move plugin to separate file
